requirements.txt
fastapi==0.120.0
sqlalchemy==2.0.36
uvicorn==0.38.0
openai==2.6.1
python-dotenv==1.1.1
python-multipart==0.0.20
pydantic==2.12.3
websockets==15.0.1
httpx==0.28.1
numpy==2.2.4
scipy==1.15.2
